Removing monads which didn't finish processing.
In the tranny program, when parsing, a monad may not finish processing because the BRAKE register went
beyond the threshold. This meant that monads with incompletely built namespaces
were then being used in the generating phase, chaos ensued. This commit fixes
that, by providing the function "kill_not_done", which removes such monads. The
tranny program now maps that function across the linked list of monads, so that
for example, "nimitzpalehuia" no longer gives hundreds of nonsense English
translations.