Add some files I forgot to Git
These include some rework in Nahuatl, according to the morphemic
analyses in the book "Introduction to Classical Nahuatl" by Andrews.
Also, the new pdf is a book I've been writing about how to program in
Tranny. The course includes beginnings of an implementation of the
Chinook Jargon (and the examples from there are also found in
languages/chinook-jargon); you might want to have a look if you don't
know how to implement a language in Tranny.