Learning
Tranny has gained a simple machine learning facility! The dictionary program
will prompt you for the words it needs (use the script in that directory) and
will fill the tranny dictionary in. Next time you re-compile the languages, it
will show up in the tranny files and will be used in parsing/generating.