Non-configurational constituency ordering is now possible
This is showcased by the rewritten Czech program (it works a little like a
conventional linked list). There is a new instruction, (check), which takes a
token and fails if that token has already been checked earlier in the parse
tree. That means that, though a verb may be either in the Thema position or in
the Rhema position, it cannot be both. So we can put (check finite-verb) in the
definition for the Czech verb.

Also included in this commit (I don't commit/push as often as I'd like, since
my networking is rubbish) is the (for) compiler macro. It takes a variable name
and a list of values, and then everything inside the macro is copied, replacing
each instance of the variable name for each value in turn.

Enjoy! Bye for now!
The Bicycle Repair Man.