Simple machine learning back in action
Look in the utils directory. There is a new program called "interrogator". If
you compile and run this, you will be able to input the vocabulary you are
asked for, instead of having to program each word into the dictionary.

It's not perfect yet, but it seems to work OK. Oh, and in the tranny source
code, there needs to be a rule defined as "Learn", which has in it the
instruction (open ...), which contains some directives to the learner. Let me
know if you find it baffling.