Trajectory
Trajectory is a new sememe that says how a verb relates to its arguments. It
has three features (for now):

 - Projective (this verb has a subject which is directly affecting the object)
     examples: I broke the mirror.

 - Reflexive (this verb has a subject which is coreferent to the object)
     examples: I can see myself.

 - Reciprocative (they are doing it to eachother)
     examples: They kissed eachother.

This has been implemented in Swahili and Ainu; both of which have a range of
valency alternating verbal affixes. Eventually, probably in the next commit, I
will show you how a language like English will handle it.

Oh, and I came across a bug meaning that the index sometimes went beyond the
end of the array. I think I've fixed it though.