More on English pronouns
Well I have tidied up the source ocde for this abit now. Hopefully that leaves
it a little more understandable. In the future I hope to set up some
remember/recall functionality, so that instead of binding a pronoun to some
variable that means "I'm a pronoun", it will instead say "recall something
that has been remembered, or will be remembered, and then refer to it".