Got rid of compiler warnings
There was a warning that an integer ws being compared to a pointer (somehow,
the code still worked okay). The offending line was tranny/parse.c:116. It is
meant to check if the index (which keeps track of how far we've gone) ever goes
past the end of the character string we're parsing.