Trajectory bug had crept in
Since the recent revert, a problem crept in that meant the trajectory variable
was not being bound in all cases. This is now fixed for English and Swahili,
as far as I can tell. This was because an old version of languages/swahili/main
and languages/swahili/verbs/verbs got copied into the working tree.