Words
A new utility in the utils/ directory is the words program. This generates all
possible words in the language you say. This might be used either for
spellchecking the sourcecode for the language, or for making a new spellchecker
(especially for languages with a high morpheme-to-word ratio!)

What inspired this is that myspell swahili accepts "ninajiona", but not
"ninajiangalia". From what I can see, both are correct.