Redone Czech aspect
In truth, verbal aspect in Czech is a little more complicated than what was
represented by the tranny source code for Czech until now. This commit redoes
the aspectual system and is much closer to the truth (but as yet it does not
work because there are so few verbs in the czech dictionary so far).
And another thing: languages/english/nouns/species needed to be added to Git.
This commit fixes that too.