Trajectory fix for Czech
I have mended the Czech way of handling trajectory.
(remember a few commits ago? d21a5eeeb3 introduced a sememe in the ontology,
that defines how a verb relates to its arguments, but I only adjusted
English and Swahili. This commit updates Czech to do the same thing.

And also, the English reciprocative pronoun "eachother" now checks that the
subject is plural. This means that a sentence like "I can see eachother" does
not get generated.