Segmentation
This commit introduces the necessary functionality to match some underlying
element to one or many orthographic representations. The use of this and the
way to do this is explained in the file languages/nahuatl/orthography/intro.

Much of Nahuatl is terribly out-of-date, which is why nothing translates to or
from it. (It refers to things in the ontology that aren't there in either
English or Swahili). I'll sort it out in the near future.