Clean compile
I have cleaned up the code to see that neither compilers (not GCC for the
source code in C, nor TC for the source code in Tranny), produce any warnings
any longer. The rationale behind this is that now it will be easy to see if
anything is out of place (type mismatches, ontological problems etc. etc.)