Massive speedup
Why was the program running so slowly? I used the GNU profiler to find out. It
turned out that one of my functions was so inefficient that it took more than
95% of the program's run-time on a typical run, so it's been rewritten. Now,
the program spends less than 0.2% of its time inside this function.

The other thing this commit addresses is a load of spelling problems with
Nahuatl. The first commit that introduced Segments was done a little naively;
This now mimics more closely what Carochi did.