Code reduction
This commit is part of an on-going attempt to reduce the amount of redundancy
in the C source code. This will make the program somewhat more efficient, and
make the code more easily understandable.

This commit introduces the file tranny/misc.c and tranny/tranny.h. misc.c
takes a lot of lines out of
  - tranny/parse.c
  - tranny/generate.c
  - tranny/attest.c
  - tranny/learn.c
It replaces that old crud with better subroutines to execute some of the most
common instructions that can be done quickly and simply.