Further reduction
Instead of forever adding loads of comparisons to the strings "seme", "rection"
"theta" etc in the bodies of the functions tranny_parse, tranny_generate etc,
these now just call tranny_binders to do it all for them. This commit is
another step in the effort to reduce redundant and duplicated code in the
project.