Correct to Swahili verbs
Until now, there was a problem where Swahili verbs would be generated with an
object prefix -- even when there was no object in the ontological
representation of the world. And the copula (with no argument) was being
selected too. Here are some examples of swahili translations of "I love"
from before this fix:

 - mimi nimekupenda .
 - nimejipenda .
 - nimevipenda .
 - mimi ni .
 - mimi si .
etc. Well, this has been fixed and "I love." now produces only sensible
translations.