HOWTOBIND
HOWTOBIND operations and the corresponding field in the monad struct define how
a string of variables is to be bound. These are boolean values ORed together
into an int. This commit fixes a bug that had crept in, which meant these flags
were cleared shortly before the the variables were being bound. So, that meant
the BLOCK flags did not have the desired effect. The BLOCK flag says: Instead
of binding this variable to a value, just delete it from the scope instead.

That was causing a subtle bug in Nahuatl because this relies on blocking some
sandhi values.