Optimisations
Profiling the tranny program revealed that a lot of time was being spent in the
functions list_get_token and list_get_list. These functions had bounds checking
to make sure you weren't requesting something past the end of the list. Another
function that was taking a lot of time was list_extend. That function extends
the memory allocated to the list, and is called if you append something to the
list and it needs more space. This commit puts a couple of things in list.h
that let you tweak these functions a bit.

Remember the segmentation that was introduced a few commits ago? Nahuatl using
this feature to support several orthographies? The tranny program now is able
to select an orthography by a command-line switch. Try these commands:

$ tranny english nahuatl orthography carochi -- I love you.
$ tranny english nahuatl orthography classical -- I love you.

This commit also does a few tweaks to the Nahuatl verb-suffices indicating
verb tense and subject's number.