Pohybná samohláska
The fleeting Czech vowel. Well, I have at last implemented that awful "feature"
of the Czech language, the vowel near the end of certain nons that sometimes
vanishes in certain cases/numbers. This is a matter of segmentation;
A consonant is a segment that may either suppress the fleeting vowel, or come
the end of the sentence, but not both. This is how it seems to work in an
actual Czech's brain, so this is how I program my computer.