Flashbulbs and context
I have introduced something called the "flashbulb". A flashbulb is a named
instance of memory, which can be called on to do various things or retrieve a
piece of information. I have used these to parse English nounphrases like the
following:
    "my mother and father",
    "my mother and my father"

They mean about the same thing, but the determiner is always present in the
first nounphrase, and may be dropped in the second. I have posited that in any
two nounphrases joined by a conjunction, the determiner of the second one may
be dropped. The phrase containing both NPs sets up a flashbulb called
"determiner", and the determiner in the first nounphrase saves some information
in it. The second nounphrase can then incorporate some definition of
"determiner" that recalls some information from the flashbulb if it exists.
Last, the flashbulb is removed.

Expect some refinement over the coming weeks; I have not played much about with
this.