Learning continues
Czech now has a Learn rule for learning nouns. Preliminary testing indicates
that it seems to work OK. The interrogator, or utils/interrogator.c, has been
altered so that it now appends a period (".") to the string you've put in.
This forces Learn to search all the way to the end of the word.