Japaneeeeese!
Tranny now has preliminary stuff for Japanese. There are some unresolved
problems with it and the vocabulary is limited to less than 10 words, but the
exciting development is the "kanjify" program in the demos/ directory. The
idea with this is that you can type a sentence in romaji, hiragana or simple
kanji, and it will use Tranny to fill the kanji into your sentence.

By-the-by:
This commit also removes some out-of-date things that were hanging around in
the demos/ folder. Also, tranny/parse.c and tranny/generate.c have reducers
now, which are little functions that can keep rules from being spawned
needlessly.