Kanjify bugfix
Until now, Kanjify was only interpreting a japanese sentence and then
generating a new one with kanji in it. Now, that's a terrible thing, because it
means that the sentence is a different one (with the same meaning and different
script). An example of this problem is that you might get "watakushi" back when
you put in "watashi" (It is a slightly different word, ne.).

This commit fixes that problem by introducing the (record ...) instruction.
This keeps track of a list of tokens, which must be matched while generating.
Ergo, you will get the same sentence back.