Generative phrase structure
This commit includes some stuff that lets phrases be generated at run-time. It
works with the (phrase) instruction, which contain any number of bars. A bar is
a thing that has (placement) information and (exec) too. The exec contains
instructions. The instructions of each bar are placed relative to each other
before being executed.