Imports
This commit introduces the imports/ directory, which will eventually take
existing computer readable dictionaries and convert them to Panini source
code. For now, there is imports/kanji, a program which takes the freely
available kanjidic (a dictionary of the Japanese symbols adapted from Chinese)
and converts it into a panini compatible format.

In other words, for each Kanji in the dictionary, this program defines the
kanji function so that it may be used by the other functions in
languages/japanese.