EDICT and KANJIDIC importers now functional
I have slaved over this hot stove for a just over a week, but the reward is
that the import folder now contains some mighty useful tools that write a
great deal of source code at compile time from the EDICT and KANJIDIC files if
they are installed on your computer. This program is nowhere near perfect yet
(for a start it only picks up nouns that are written with a single kanji) but
it's a start and much more jap vocab is now available for use with libpanini.