Algonquian
There are now a few words of algonquian in here, and also all necessary
segments have their reflexes to some dialect of Ojibwe. In the future, this
will mean that, in theory, as data gets gathered for the various Algonquian
languages, it will be available for other Algonquian languages. If I arrange
for Proto-Algonquian to get reconstructed at compile-time, and then its
descendants get "reflected" at run-time, then if a word gets learned in Munsee
and Blackfoot, then it will be readily available also in Fox, Cree, Menominee
etc. Whether this will work at all remains to be seen.