Fixed the ekan importer.
I think the ekan importer (it reads the EDICT and KANJIDIC dictionaries and
generates panini source code) was broken until now. Now it's working so that's
a bit better, still not perfect.

And also, there was a bug in vars/bind.c which was causing a segfault for some
reason, and that's also been tied up.