Abstraction overhaul and new features
There are a new batch of functions beginning with panini_, which are (for now)
essentially a layer of abstraction over the old tranny_* routines. They expose
the existing functionality in a more convenient way,  and also expose some new
features like spellchecking.

Have a look, using monad_map to crawl across the linked list is beginning to
deprecate