PP learner now learns to conjugate its verbs :-)
Until now, whenever the pp learner learned a language, it recognised a verbstem
and some tag on the end. That thing, of course, encodes TAM and the subject's
Phi-features, but they were ignored until now.

Now, if you run the pp program to learn Spanish for example, and install it,
and translate "I kiss ." you get:
 - beeso
 - beso
 - beo
 - bso
 - bo
 - beseso
 - besso
 - beseeso
 - beseo
Not perfect (I need to make some sort of attester for that), but a lot better
than it used to be.