Multistage machine learning now reasonable.
This is exemplified in the languages/spanish directory, which learns spanish
from the examples in data.txt. First we run stage0, which determines how basic
conjugation works in this language. Then we run stage1, which learns some verbs
and finally, stage2 learns some pronouns. Later of course, we may put other
stages in place to learn other parts of speech and other syntactic forms etc.