The Golden Rule of Programming
I don't know if this is an actual thing, but it should be. The Golden Rule of
Programming says that if you are writing the same, or very similar, code
several times, then you're doing it wrong.

This commit begins to address this by introducing a function panini_despatch,
which will soon eclipse all of the following:
 - tranny_parse
 - tranny_generate
 - tranny_learn
 - tranny_attest
because these functions are very slightly different, the idea is that the new
function will be parameterised to make for differences like "Are we appending
strings to the outtext string or are we reading them in from the intext
string?"