Flashcards
Once we have had a computer learn a language, why not use it for humans'
benefit? That is the point of the Panini Project. You do not only get yet
another machine translation program, you get flashcards for the memorisation
app you've got on your smartphone, and you get a human readable dictionary
(coming soon), and many other goodies.

This commit makes a start on the flashcards program, found in
./demos/flashcards/.