World
World is a new thing in languages/world, which describes the world in more
detail than ontology.csv does. The point of this is to resolve ambiguities such
as "He's wearing a rubber" (which interpretation of "rubber" should we choose
here?), and incidentally the flashcards program uses this to categorise the
vocabulary.