The Bible is now used as a corpus of text
The bible is widely available for free, and consists of many sentences that
have been carefully translated from one language to another by hand. This is
conveniently usable by my machine learning tools.

Whether or not you want to use this feature at compile time is configurable in
the environment variables set in ./build.sh.