Learning has been slightly rearranged
Now, instead of the single function panini_learn which learned syntax and
vocabulary and other morphemes, this functionality has been split into two
separate functions:
 - panini_learnpp for learning and parameterising syntax,
 - panini_learnvocab for learning morphemes

The disadvantage is that a language module must first contain enough words to
start learning any grammar, and then be rebuilt to include the grammar learned.
But on the other hand, the resulting programs are much faster and consume much
less memory. The whole project now compiles on my computer with 4GB of memory
(and probably requires much much less).