Fine tuning
This commit fine tunes the build process a little bit. It is now possible to
configure which languages you're bothered about building, and also you can use
clang as a compiler for most things if you wish to.

Also, some changes have been made to tools/ekan/edict.c, so that it now also
can find verbs in the dictionary (for now only ichidan, but later godan verbs
and i-keiyoushi will be searched for).