English bootstrap
I have just made a change to how the English words are done. Until now, each
English word that was also a semantic prime was defined as such with (lit ...).
The problem is that (lit ) does a case sensitive match, which means that, for
example, "earth" would not match "Earth". This commit introduces a program
tools/learn/bootstrap (which is only to be used with English!).

What this program does: It takes all the semantic primes which are also English
words. These are listed in languages/english/bootstrap.txt. Each one is
examined, resulting in a dictionary definition that can be used by panini. Then
the english module is recompiled to include these definitions.

The hope is that it will be much much quicker now to add a semantic prime (and
the english word in one go). And that it should be possible from now on to
process English text with proper capitalisation.