Unihan
When creating large projects, it is prudent and sensible to try and not
reinvent any wheels. This commit introduces a new importer, unihan.

Unihan is the name of a whole load of data published by the Unicode
Consortium, which documents various Han characters (these are used to write
Chinese, Japanese and more). Unihan is also the name of the importer that's
part of the Panini Project which takes this data and presents it in a form
that may be used by Panini targets.