A lot has changed.
This commit introduces an interface to wiktionary which has eclipsed WordNet as
the way to bootstrap the English language knowledge base.

Also new is an interface to Unihan, the excellent resource on CJK glyphs
provided by the Unicode consortium. Hopefully, this will eventually be used to
make a target supporting the Chinese languages.