Start on automatic code-generation
The DRY principle, or "Don't Repeat Yourself", means that a piece of code
should ideally only be written once if it only needs to be written once. As
language features are often shared among many various languages, this principle
can contribute to unifying the various targets we've got so far, and might help
in making new ones more speedily, too.

For now, we've got a thing in tools/wals/ which interrogated the World Atlas of
Language Structures for features that a language has. If it is found to have
prepositions, then a particular source file may be copied in to the src tree of
that particular target. Ditto if the language is found to have postpositions.

So, we need to look at getting existing targets to use this feature where
applicable; this will reduce the effort needed to maintain them. We should also
look at adding more and more language features into this tool. Theoretically,
that should not be too hard ...

More to follow.