Improvements to the build system
the build.sh script has been simplified. A few simple for loops now
delegate the compilation, downloading, building etc to other scripts in
subdirectories.

More importantly, the Autogen feature has been given an overhaul. The old
programs done in bash have been retired and replaced with a new system that
has been made in python. (AFAICT) it now produces something that can parse
simple Chinese sentences (except none of the words are there yet, but that's
the easy part). I'll put it to the test properly before the next commit.